* adding error to the job summary  when sast failed *adding httpcomponent dependency *organize the plugin libraries *remove unnecessary files (includes th ui template) *ready for release

+++ src/main/java/com/cx/plugin/utils/CxFileChecker.java
@@ -1,40 +1,37 @@
-package com.cx.plugin.dto;
+package com.cx.plugin.utils;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
 import java.io.File;
 import java.io.FilenameFilter;
 
 /**
  * Created by Galn on 28/02/2017.
  */
-public class FileChecker {//TODO - Dor - how make the calss static- ther should be one way
-    public static final Logger log = LoggerFactory.getLogger(FileChecker.class);
+public abstract class CxFileChecker {
+    public static final Logger log = LoggerFactory.getLogger(CxFileChecker.class);
 
     public static void deleteFile(String folder, String prefix){
 
         GenericPrefixFilter filter = new GenericPrefixFilter(prefix);
         File dir = new File(folder);
 
-        //list out all the file name with .txt extension
+        //list out all the file name with prefix
         String[] list = dir.list(filter);
 
-        if (list.length == 0) return;
+        if (list==null || list.length == 0) return;
 
         File fileDelete;
 
         for (String file : list){
-            String temp = new StringBuffer(folder)
-                    .append(File.separator)
-                    .append(file).toString();
+            String temp = folder + File.separator + file;
             fileDelete = new File(temp);
             boolean isDeleted = fileDelete.delete();
             log.warn("file : " + temp + " is deleted : " + isDeleted);
         }
     }
 
-    //inner class, generic extension filter
+    //inner class, generic prefix filter
     public static class GenericPrefixFilter implements FilenameFilter {
 
         private String prefix;
